requirements.txt
matplotlib==3.5.2
numpy==1.22.3
pandas==1.4.2
Pillow==9.1.0
plotly==5.7.0
requests==2.27.1
seaborn==0.11.2
streamlit==1.8.1
streamlit_lottie==0.0.3
